feature: create cookie policy and cookie consent
