+++ components/common/FooterComponent.tsx
@@ -19,7 +19,7 @@ const navigation = [
 
 const FooterComponent = () => {
   return (
-    <footer className='bg-white relative inset-x-0 bottom-0 z-50'>
+    <footer className='bg-white relative inset-x-0 bottom-0 z-30'>
       <div className='container px-6 py-12 mx-auto'>
         <div className='grid grid-cols-1 gap-6 sm:grid-cols-2 sm:gap-y-10 lg:grid-cols-4'>
           <SubscribeForm />
