+++ components/common/HeaderComponent.tsx
@@ -17,7 +17,7 @@ const HeaderComponent = () => {
   const [mobileMenuOpen, setMobileMenuOpen] = useState(false)
 
   return (
-    <header className='bg-white relative inset-x-0 top-0 z-50'>
+    <header className='bg-white relative inset-x-0 top-0 z-30'>
       <nav
         className='container flex items-center justify-between px-6 py-6 mx-auto'
         aria-label='Global'
